Dev/raquela/bug fixes 0416 (#62)
* - Adding helpers to get selected ProjectItem and selected ProjectItem's Project.
This also fixes error when invoking "Managing client side libraries" Project menu option and there are no projects selected.
- Adding check to delete files only if exists.

+++ src/LibraryManager.Build/Contracts/HostInteraction.cs
@@ -66,9 +66,11 @@ namespace Microsoft.Web.LibraryManager.Build
 
                 try
                 {
-                    File.Delete(absoluteFile);
-
-                    Logger.Log(string.Format(Resources.Text.FileDeleted, relativeFilePath), LogLevel.Operation);
+                    if (File.Exists(absoluteFile))
+                    {
+                        File.Delete(absoluteFile);
+                        Logger.Log(string.Format(Resources.Text.FileDeleted, relativeFilePath), LogLevel.Operation);
+                    }
                 }
                 catch (Exception)
                 {
